backend/pyproject.toml
[tool.poetry]
name = "geoplotter-be"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "geoplotter_be"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.1"
uvicorn = "^0.21.1"
pydantic = "^1.10.7"
beanie = "^1.18.0"
bcrypt = "^4.0.1"
pyjwt = "^2.6.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
